Fixing connect->disconnect->connect issues for ConsumerService. Move definition for the designated classes to specified files.
