+++ computationalLib/include/computationalLib/network_shared/amqpCommon.h
@@ -98,121 +98,8 @@ class MyHandler : public AMQP::LibBoostAsioHandler {
   bool connected_;
 };
 
-/**
- * @brief Service that allow to Publish messages to multiple queues
- */
- //todo set new queue(connstring)
-class amqpPublisherService {
- public:
-
-  /**
-   * @details Will declare all queues in provide array
-   * @param connection_string
-   * @param queues
-   */
-  amqpPublisherService(const std::string &connection_string,
-                       const std::vector<std::string> &queues = {});
-
-  /**
-   * @brief delted i's queue
-   * @param i
-   * @throws shared::outOfRange
-   */
-  void RemoveQueue(size_t i);
-
-  /**
-   * Adds new queue
-   * @param queue
-   * @param create
-   */
-  void AddQueue(const std::string &queue,
-                bool create);
-
-  /**
-   * @brief publishe message to i's queue
-   * @param message
-   * @param i
-   * @throws shared::outOfRange
-   */
-  void Publish(EnvelopePtr message, size_t i);
-
-  void EndLoop();
-
-  void RestartLoop();
-
-  ~amqpPublisherService();
-
-  bool IsConnected() const;
-
- private:
-  boost::asio::io_service m_service;
-  std::unique_ptr<boost::asio::io_service::work> m_work;
-
-  MyHandler m_handler;
-  AMQP::TcpConnection m_connection;
-  AMQP::TcpChannel m_channel;
-  std::vector<std::string> m_queues;
-  std::thread m_serviceThread;
-
-  static inline const std::string defaultExhc = "testexch";//todo delete
-
-};
-
-/**
- * @brief Service that allow to start and end reading event Loop from specified queue
- */
-class amqpConsumerService {
- public:
 
 
-  /**
-   *
-   */
-  amqpConsumerService();
-  /**
-   * @brief
-   * @param connection_string
-   * @param queue_name
-   */
-  amqpConsumerService(const std::string &connection_string,
-                      const std::string &queue_name);
 
-
-  /**
-   * @brief
-   * @param connection_string
-   * @param queue_name
-   */
-  void SetParameters(const std::string &connection_string,
-                    const std::string &queue_name);
-  /**
-   * @brief Starts event Loop to read from queue1
-   */
-  void Connect();
-
-  /**
-   * @brief Ends event Loop()
-   */
-  void Disconnect();
-
-  ~amqpConsumerService();
-
-  bool IsConnected() const;
-
- private:
-
-  void Reconnect();
-
-  boost::asio::io_service m_service;
-  std::unique_ptr<boost::asio::io_service::work> m_work;
-
-  std::unique_ptr<MyHandler> m_handler;
-  std::unique_ptr<AMQP::TcpConnection> m_connection;
-  std::unique_ptr<AMQP::TcpChannel> m_channel;
-
-  std::string m_queue;
-  std::string m_c_string;
-  std::thread m_serviceThread;
-};
 
 }
